System.Directory.Layout

Description

Language to express directory layouts

Layout declaration

data DL a

Representation of directory layouts

Invariants:

F second argument is never D _ _ _ or F _ _ _ itself

F third argument is never T _ _

D second argument is never T _ _

D third argument is never T _ _

type Layout = DL ()

Type synonym to save some acrobatics

file :: FilePath -> Text -> Layout

Declare file with specified contents

file_ :: FilePath -> Layout

Declare empty file

directory :: FilePath -> Layout -> Layout

Declare directory with specified listing

directory_ :: FilePath -> Layout

Declare empty directory

Layout construction

fromDirectory :: FilePath -> IO (Either IOException Layout)

Create layout from directory

Canonicalizes path before traversing, generally understands only regular files and directories and ignores anything else it could not understand

Layout traverses

make

Arguments

:: Layout

-> FilePath
Root directory

-> IO [LayoutException]
List of warnings

Make layout as specified

For example, suppose you are in an empty directory

 % tree
 .

and you've written simple layout:

 layout = do
   directory "baz" $
     file_ "twey"
   directory "foo" $ do
     directory "bar" $ do
       file_ "quuz"
       file_ "tatata"
     file_ "quux"

then running it should result in this directory tree:

 % tree
 .
 ├── baz
 │   └── twey
 └── foo
     ├── bar
     │   ├── quuz
     │   └── tatata
     └── quux

check

Arguments

:: Layout

-> FilePath
Root directory

-> IO [LayoutException]
List of failures

Check directory layout agrees with specified one

For example, suppose there is a tree:

 % tree
 .
 ├── baz
 │   └── twey
 └── foo
     ├── bar
     │   ├── quuz
     │   └── tatata
     └── quux

then you can write:

 layout = do
   directory "baz" $
     file_ "twey"
   directory "foo" $ do
     directory "bar" $ do
       file_ "quuz"
       file_ "tatata"
     file_ "quux"

and running check layout "." should result in []

Errors

data LayoutException

Information about cought exceptions in various routines

Constructors

CD IOErrorType FilePath
createDirectory exceptions

CF IOErrorType FilePath
createFile eceptions

FE IOErrorType FilePath
fileExists eceptions

DE IOErrorType FilePath
directoryExists eceptions

RF IOErrorType FilePath Text
readFile eceptions