System.Directory.Layout.Lens

Description

Control.Lens based extractors for DL

Documentation

>>> :set -XOverloadedStrings
>>> import           Control.Lens
>>> let layout = F "foo" (T "not empty" ()) (D "bar" (F "baz" (E ()) (F "quux" (T "something" ()) (E ()))) (F "swaks" (E ()) (E ())))

text :: Prism Layout Layout Text Text

Get Text out of the current Layout (if possible)

>>> layout ^? text
Nothing
>>> layout ^? file "foo" . text
Just "not empty"
>>> layout ^? directory "bar" . file "quux" . text
Just "something"

file :: FilePath -> IndexedTraversal' FilePath Layout Layout

Look into the file in the current Layout (if possible)

>>> layout ^? file "biz"
Nothing
>>> layout ^? file "swaks"
Just (E ())
>>> layout ^? directory "bar" . file "baz"
Just (E ())

directory :: FilePath -> IndexedTraversal' FilePath Layout Layout

Go into the directory in the current Layout (if possible)

>>> layout ^? directory "foo"
Nothing
>>> layout ^? directory "bar"
Just (F "baz" (E ()) (F "quux" (T "something" ()) (E ())))