System.Directory.Layout.Make

Description

Make layout as specified

For example, suppose you are in an empty directory

 % tree
 .

and you've written simple layout:

 layout = do
   directory "baz" $
     file_ "twey"
   directory "foo" $ do
     directory "bar" $ do
       file_ "quuz"
       file_ "tatata"
     file_ "quux"

then running it should result in this directory tree:

 % tree
 .
 ├── baz
 │   └── twey
 └── foo
     ├── bar
     │   ├── quuz
     │   └── tatata
     └── quux

Documentation

data DLMakeWarning

Data type representing various warnings that may occur while infecting directory layout

Constructors

FileDoesExist FilePath

DirectoryDoesExist FilePath

make

Arguments

:: Layout

-> FilePath
Root directory

-> IO [DLMakeWarning]
List of warnings

Infect file layout with stuff from script