System.Directory.Layout

Layout declaration

data DL f

Abstract data type representing directory tree is nice

type Layout = DL ()

But type synonym is nicer

file :: FilePath -> Text -> Layout

Declare file with specified contents

file_ :: FilePath -> Layout

Declare empty file

directory :: FilePath -> Layout -> Layout

Declare directory with specified listing

directory_ :: FilePath -> Layout

Declare empty directory

Layout construction

data DLMakeWarning

Data type representing various warnings that may occur while infecting directory layout

Constructors

FileDoesExist FilePath

DirectoryDoesExist FilePath

make

Arguments

:: Layout

-> FilePath
Root directory

-> IO [DLMakeWarning]
List of warnings

Infect file layout with stuff from script

Layout verification

data DLCheckFailure

Data type representing various failures that may occur while checking directory layout

Constructors

FileDoesNotExist FilePath

FileWrongContents FilePath Text

DirectoryDoesNotExist FilePath

check

Arguments

:: Layout

-> FilePath
Root directory

-> IO [DLCheckFailure]
List of failures

Check directory layout corresponds to specified one